+++ spec/integration/valid_examples_spec.rb
@@ -15,7 +15,6 @@ RSpec.describe 'valid examples' do
     image-url
     image
     index
-    min
     nomedia
     profile
     required