+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 # 0.2.0
 
-* Normalize case of properties
+* Normalize case of properties.
+* Implement required property and vertical validation.